Add adjustable goal metric/measurment to AddHabit, create WeekView
